test: fold oauth credential scoping tests into the mapped pre-call suite
